refactor(desktop): 拆出 skills 与 rules 设置面板
将 SkillsSettingsPanel、RulesSettingsPanel 及辅助函数迁入 settings/panels/。
